add lighthouse-lr-background
